+++ src/GafferSceneUI/VisualiserTool.cpp
@@ -287,6 +287,122 @@ std::string const g_vertexLabelShaderFragSource
 );
 
 //-----------------------------------------------------------------------------
+// Vector shader
+//-----------------------------------------------------------------------------
+
+struct UniformBlockVectorShader
+{
+	alignas( 16 ) Imath::M44f o2v;
+	alignas( 16 ) Imath::M44f n2v;
+	alignas( 16 ) Imath::M44f v2c;
+	alignas( 16 ) Imath::M44f o2c;
+	alignas( 16 ) Imath::Color3f color;
+	alignas( 4 ) float scale;
+};
+
+#define UNIFORM_BLOCK_VECTOR_GLSL_SOURCE \
+	"layout( std140, row_major ) uniform UniformBlock\n" \
+	"{\n" \
+	"   mat4 o2v;\n" \
+	"   mat4 n2v;\n" \
+	"   mat4 v2c;\n" \
+	"   mat4 o2c;\n" \
+	"   vec3 color;\n" \
+	"   float scale;\n" \
+	"} uniforms;\n"
+
+#define ATTRIB_GLSL_LOCATION_VS 1
+
+#define ATTRIB_VECTOR_GLSL_SOURCE \
+	"layout( location = " BOOST_PP_STRINGIZE( ATTRIB_GLSL_LOCATION_PS ) " ) in vec3 ps;\n" \
+	"layout( location = " BOOST_PP_STRINGIZE( ATTRIB_GLSL_LOCATION_VS ) " ) in vec3 vs;\n"
+
+// Opengl vertex shader code (point format)
+
+const std::string g_vectorShaderVertSourcePoint
+(
+	"#version 330\n"
+
+	UNIFORM_BLOCK_VECTOR_GLSL_SOURCE
+
+	ATTRIB_VECTOR_GLSL_SOURCE
+
+	"void main()\n"
+	"{\n"
+	"   vec3 position = ps;\n"
+
+	"   if( gl_VertexID == 1 )\n"
+	"   {\n"
+	"       position = vs;\n"
+	"   }\n"
+
+	"   gl_Position = vec4( position, 1.0 ) * uniforms.o2c;\n"
+	"}\n"
+);
+
+// Opengl vertex shader code (vector format)
+
+const std::string g_vectorShaderVertSourceVector
+(
+	"#version 330\n"
+
+	UNIFORM_BLOCK_VECTOR_GLSL_SOURCE
+
+	ATTRIB_VECTOR_GLSL_SOURCE
+
+	"void main()\n"
+	"{\n"
+	"   vec3 position = ps;\n"
+
+	"   if( gl_VertexID == 1 )\n"
+	"   {\n"
+	"       position += vs * uniforms.scale;"
+	"   }\n"
+
+	"   gl_Position = vec4( position, 1.0 ) * uniforms.o2c;\n"
+	"}\n"
+);
+
+// Opengl vertex shader code (bivector format)
+
+const std::string g_vectorShaderVertSourceBivector
+(
+	"#version 330\n"
+
+	UNIFORM_BLOCK_VECTOR_GLSL_SOURCE
+
+	ATTRIB_VECTOR_GLSL_SOURCE
+
+	"void main()\n"
+	"{\n"
+	"   vec4 position = vec4( ps, 1.0 ) * uniforms.o2v;\n"
+
+	"   if( gl_VertexID == 1 )\n"
+	"   {\n"
+	"       position.xyz += normalize( vs * mat3( uniforms.n2v ) ) * ( uniforms.scale * length( vs ) );\n"
+	"   }\n"
+
+	"   gl_Position = position * uniforms.v2c;\n"
+	"}\n"
+);
+
+// Opengl fragment shader code
+
+std::string const g_vectorShaderFragSource
+(
+	"#version 330\n"
+
+	UNIFORM_BLOCK_VECTOR_GLSL_SOURCE
+
+	"layout( location = 0 ) out vec4 cs;\n"
+
+	"void main()\n"
+	"{\n"
+	"   cs = vec4( uniforms.color, 1.0 );\n"
+	"}\n"
+);
+
+//-----------------------------------------------------------------------------
 // Helper Methods
 //-----------------------------------------------------------------------------
 
@@ -420,6 +536,7 @@ class VisualiserGadget : public Gadget
 			if( layer == Gadget::Layer::MidFront )
 			{
 				renderColorVisualiser( viewportGadget, mode );
+				renderVectorVisualiser( viewportGadget, mode );
 			}
 
 			else if( layer == Gadget::Layer::Front )
@@ -659,9 +776,13 @@ class VisualiserGadget : public Gadget
 
 				ConstDataPtr vData = vIt->second.data;
 
-				if( mode == VisualiserTool::Mode::Auto && vData->typeId() == IntVectorDataTypeId )
+				if(
+					mode == VisualiserTool::Mode::Auto && (
+						vData->typeId() == IntVectorDataTypeId ||  // Will be handled by `renderVertexLabelValue()` instead.
+						vData->typeId() == V3fVectorDataTypeId  // Will be handled by `renderVectorVisualiser()` instead.
+					)
+				)
 				{
-					// Will be handled by `renderVertexLabelValue()` instead.
 					continue;
 				}
 
@@ -834,7 +955,12 @@ class VisualiserGadget : public Gadget
 
 			const VisualiserTool::CursorValue value = m_tool->cursorValue();
 
-			if( mode == VisualiserTool::Mode::Auto && std::holds_alternative<int>( value ) )
+			if(
+				mode == VisualiserTool::Mode::Auto && (
+					std::holds_alternative<int>( value ) ||
+					std::holds_alternative<V3f>( value )
+				)
+			)
 			{
 				return;
 			}
@@ -1007,6 +1133,12 @@ class VisualiserGadget : public Gadget
 
 					if( !vData )
 					{
+						continue;
+					}
+
+					if( mode == VisualiserTool::Mode::Auto && vData->typeId() == V3fVectorDataTypeId )
+					{
+						// Will be handled by `renderVectorVisualiser()` instead.
 						continue;
 					}
 
@@ -1348,6 +1480,237 @@ class VisualiserGadget : public Gadget
 			m_cursorVertexValue = cursorVertexValue;
 		}
 
+		void renderVectorVisualiser( const ViewportGadget *viewportGadget, VisualiserTool::Mode mode ) const
+		{
+			const std::string name = primitiveVariableFromDataName( m_tool->dataNamePlug()->getValue() );
+			if( name.empty() || mode != VisualiserTool::Mode::Auto )
+			{
+				return;
+			}
+
+			buildShader( m_vectorShaderPoint, g_vectorShaderVertSourcePoint, g_vectorShaderFragSource );
+			buildShader( m_vectorShaderVector, g_vectorShaderVertSourceVector, g_vectorShaderFragSource );
+			buildShader( m_vectorShaderBivector, g_vectorShaderVertSourceBivector, g_vectorShaderFragSource );
+
+			if( !m_vectorShaderPoint || !m_vectorShaderVector || !m_vectorShaderBivector )
+			{
+				return;
+			}
+
+			// Get the cached converter from IECoreGL, this is used to convert primitive
+			// variable data to opengl buffers which will be shared with the IECoreGL renderer
+			IECoreGL::CachedConverter *converter = IECoreGL::CachedConverter::defaultCachedConverter();
+
+			GLint uniformBinding;
+			glGetIntegerv( GL_UNIFORM_BUFFER_BINDING, &uniformBinding );
+
+			if( !m_vectorUniformBuffer )
+			{
+				GLuint buffer = 0u;
+				glGenBuffers( 1, &buffer );
+				glBindBuffer( GL_UNIFORM_BUFFER, buffer );
+				glBufferData( GL_UNIFORM_BUFFER, sizeof( UniformBlockVectorShader ), 0, GL_DYNAMIC_DRAW );
+				m_vectorUniformBuffer.reset( new IECoreGL::Buffer( buffer ) );
+			}
+
+			glBindBufferBase( GL_UNIFORM_BUFFER, g_uniformBlockBindingIndex, m_vectorUniformBuffer->buffer() );
+
+			UniformBlockVectorShader uniforms;
+			uniforms.color = Color3f( 1.f );
+			uniforms.scale = 1.f;
+
+			// Get the world to view and view to clip space matrices
+			const M44f w2v = viewportGadget->getCameraTransform().gjInverse();
+			glGetFloatv( GL_PROJECTION_MATRIX, uniforms.v2c.getValue() );
+
+			// Set OpenGL state
+			GLfloat lineWidth;
+			glGetFloatv( GL_LINE_WIDTH, &lineWidth );
+			glLineWidth( 1.f );
+
+			const GLboolean depthEnabled = glIsEnabled( GL_DEPTH_TEST );
+			if( !depthEnabled )
+			{
+				glEnable( GL_DEPTH_TEST );
+			}
+
+			GLboolean depthWriteEnabled;
+			glGetBooleanv( GL_DEPTH_WRITEMASK, &depthWriteEnabled );
+			if( depthWriteEnabled )
+			{
+				glDepthMask( GL_FALSE );
+			}
+
+			GLboolean lineSmooth;
+			glGetBooleanv( GL_LINE_SMOOTH, &lineSmooth );
+			if( lineSmooth )
+			{
+				glDisable( GL_LINE_SMOOTH );
+			}
+
+			const GLboolean blendEnabled = glIsEnabled( GL_BLEND );
+			if( blendEnabled )
+			{
+				glDisable( GL_BLEND );
+			}
+
+			// Store current shader program to be restored after drawing.
+			// We set the shader program for drawing vectors when we know
+			// the interpretation of the visualised data, which may
+			// be different per object.
+			GLint shaderProgram;
+			glGetIntegerv( GL_CURRENT_PROGRAM, &shaderProgram );
+			std::optional<GLint> currentShaderProgram;
+
+			// Set OpenGL vertex attribute array state
+			GLint arrayBinding;
+			glGetIntegerv( GL_ARRAY_BUFFER_BINDING, &arrayBinding );
+
+			glPushClientAttrib( GL_CLIENT_VERTEX_ARRAY_BIT );
+
+			glVertexAttribDivisor( ATTRIB_GLSL_LOCATION_PS, 1 );
+			glEnableVertexAttribArray( ATTRIB_GLSL_LOCATION_PS );
+			glVertexAttribDivisor( ATTRIB_GLSL_LOCATION_VS, 1 );
+			glEnableVertexAttribArray( ATTRIB_GLSL_LOCATION_VS );
+
+			// Loop through the current selection
+			for( const auto &location : m_tool->selection() )
+			{
+				ScenePlug::PathScope scope( &location.context(), &location.path() );
+
+				// Check path exists
+				if( !location.scene().existsPlug()->getValue() )
+				{
+					continue;
+				}
+
+				// Extract primitive
+				auto primitive = runTimeCast<const Primitive>( location.scene().objectPlug()->getValue() );
+				if( !primitive )
+				{
+					continue;
+				}
+
+				// Make sure we have "P" data and it is the correct type.
+				const auto pIt = primitive->variables.find( g_pName );
+				if( pIt == primitive->variables.end() )
+				{
+					continue;
+				}
+
+				auto pData = runTimeCast<const V3fVectorData>( pIt->second.data );
+				if( !pData )
+				{
+					continue;
+				}
+
+				// Find named vertex attribute
+				// NOTE : Conversion to IECoreGL mesh may generate vertex attributes (eg. "N")
+				// so check named primitive variable exists on IECore mesh primitive.
+				const auto vIt = primitive->variables.find( name );
+				if( vIt == primitive->variables.end() )
+				{
+					continue;
+				}
+
+				auto vData = runTimeCast<const V3fVectorData>( vIt->second.data );
+				if( !vData )
+				{
+					// Will be handled by `renderColorVisualiser()` or `renderVertexLabelValue()` instead.
+					continue;
+				}
+
+				// Retrieve cached IECoreGL primitive
+				auto primitiveGL = runTimeCast<const IECoreGL::Primitive>( converter->convert( primitive.get() ) );
+				if( !primitiveGL )
+				{
+					continue;
+				}
+
+				IECoreGL::ConstBufferPtr pBuffer = primitiveGL->getVertexBuffer( g_pName );
+				if( !pBuffer )
+				{
+					continue;
+				}
+
+				IECoreGL::ConstBufferPtr vBuffer = primitiveGL->getVertexBuffer( name );
+				if( !vBuffer )
+				{
+					continue;
+				}
+
+				GLint vDataProgram;
+				switch( vData->getInterpretation() )
+				{
+					case GeometricData::Interpretation::Point :
+						vDataProgram = m_vectorShaderPoint->program();
+						break;
+					case GeometricData::Interpretation::Normal :
+						vDataProgram = m_vectorShaderBivector->program();
+						break;
+					default :
+						vDataProgram = m_vectorShaderVector->program();
+				}
+				if( !currentShaderProgram || currentShaderProgram.value() != vDataProgram )
+				{
+					glUseProgram( vDataProgram );
+					currentShaderProgram = vDataProgram;
+				}
+
+				// Get the object to world transform
+				M44f o2w;
+				ScenePlug::ScenePath path( location.path() );
+				while( !path.empty() )
+				{
+					scope.setPath( &path );
+					o2w = o2w * location.scene().transformPlug()->getValue();
+					path.pop_back();
+				}
+
+				// Compute object/normal to view and object to clip matrices
+				uniforms.o2v = o2w * w2v;
+				uniforms.n2v = ( uniforms.o2v.gjInverse() ).transpose();
+				uniforms.o2c = uniforms.o2v * uniforms.v2c;
+
+				// Upload OpenGL uniform block data
+				glBufferData( GL_UNIFORM_BUFFER, sizeof( UniformBlockVectorShader ), &uniforms, GL_DYNAMIC_DRAW );
+
+				// Instance a line segment for each element of vector data
+				glBindBuffer( GL_ARRAY_BUFFER, pBuffer->buffer() );
+				glVertexAttribPointer( ATTRIB_GLSL_LOCATION_PS, 3, GL_FLOAT, GL_FALSE, 0, nullptr );
+				glBindBuffer( GL_ARRAY_BUFFER, vBuffer->buffer() );
+				glVertexAttribPointer( ATTRIB_GLSL_LOCATION_VS, 3, GL_FLOAT, GL_FALSE, 0, nullptr );
+				glDrawArraysInstanced( GL_LINES, 0, 2, static_cast<GLsizei>( primitiveGL->getVertexCount() ) );
+
+			}
+
+			// Restore OpenGL state
+
+			glPopClientAttrib();
+			glBindBuffer( GL_ARRAY_BUFFER, arrayBinding );
+			glBindBuffer( GL_UNIFORM_BUFFER, uniformBinding );
+
+			glLineWidth( lineWidth );
+
+			if( lineSmooth )
+			{
+				glEnable( GL_LINE_SMOOTH );
+			}
+			if( blendEnabled )
+			{
+				glEnable( GL_BLEND );
+			}
+			if( !depthEnabled )
+			{
+				glDisable( GL_DEPTH_TEST );
+			}
+			if( depthWriteEnabled )
+			{
+				glDepthMask( GL_TRUE );
+			}
+			glUseProgram( shaderProgram );
+		}
+
 		VisualiserTool::CursorValue cursorVertexValue() const
 		{
 			return m_cursorVertexValue;
@@ -1358,6 +1721,10 @@ class VisualiserGadget : public Gadget
 		mutable IECoreGL::ConstBufferPtr m_colorUniformBuffer;
 		mutable IECoreGL::ConstShaderPtr m_vertexLabelShader;
 		mutable IECoreGL::ConstBufferPtr m_vertexLabelUniformBuffer;
+		mutable IECoreGL::ConstShaderPtr m_vectorShaderPoint;
+		mutable IECoreGL::ConstShaderPtr m_vectorShaderVector;
+		mutable IECoreGL::ConstShaderPtr m_vectorShaderBivector;
+		mutable IECoreGL::ConstBufferPtr m_vectorUniformBuffer;
 
 		mutable IECoreGL::ConstBufferPtr m_vertexLabelStorageBuffer;
 		mutable std::size_t m_vertexLabelStorageCapacity;